[Android] Style: data/ -> data/db/ 로 이동
